more any mention of core group
