--- openmls/src/group/core_group/mod.rs
@@ -1,206 +1,0 @@
-//! ### Don't Panic!
-//!
-//! Functions in this module should never panic. However, if there is a bug in
-//! the implementation, a function will return an unrecoverable `LibraryError`.
-//! This means that some functions that are not expected to fail and throw an
-//! error, will still return a `Result` since they may throw a `LibraryError`.
-
-// Private
-pub(super) mod new_from_welcome;
-
-// Crate
-pub(crate) mod create_commit_params;
-pub(crate) mod new_from_external_init;
-pub(crate) mod past_secrets;
-pub(crate) mod process;
-pub(crate) mod proposals;
-
-use openmls_traits::signatures::Signer;
-use serde::{Deserialize, Serialize};
-
-use self::past_secrets::MessageSecretsStore;
-
-use super::{
-    errors::{CreateAddProposalError, ValidationError},
-    public_group::PublicGroup,
-};
-
-use crate::{
-    binary_tree::array_representation::LeafNodeIndex,
-    credentials::*,
-    error::LibraryError,
-    framing::{mls_auth_content::AuthenticatedContent, *},
-    group::*,
-    key_packages::*,
-    messages::{
-        group_info::{GroupInfo, VerifiableGroupInfo},
-        proposals::*,
-        *,
-    },
-    schedule::{psk::store::ResumptionPskStore, *},
-    treesync::{node::encryption_keys::EncryptionKeyPair, *},
-};
-
-#[cfg(test)]
-use super::errors::CreateGroupContextExtProposalError;
-#[cfg(test)]
-use crate::treesync::node::leaf_node::TreePosition;
-
-#[derive(Debug)]
-pub(crate) struct CreateCommitResult {
-    pub(crate) commit: AuthenticatedContent,
-    pub(crate) welcome_option: Option<Welcome>,
-    pub(crate) staged_commit: StagedCommit,
-    pub(crate) group_info: Option<GroupInfo>,
-}
-
-/// A member in the group is identified by this [`Member`] struct.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Member {
-    /// The member's leaf index in the ratchet tree.
-    pub index: LeafNodeIndex,
-    /// The member's credential.
-    pub credential: Credential,
-    /// The member's public HPHKE encryption key.
-    pub encryption_key: Vec<u8>,
-    /// The member's public signature key.
-    pub signature_key: Vec<u8>,
-}
-
-impl Member {
-    /// Create new member.
-    pub fn new(
-        index: LeafNodeIndex,
-        encryption_key: Vec<u8>,
-        signature_key: Vec<u8>,
-        credential: Credential,
-    ) -> Self {
-        Self {
-            index,
-            encryption_key,
-            signature_key,
-            credential,
-        }
-    }
-}
-
-impl MlsGroup {
-    // === Create handshake messages ===
-    // TODO: share functionality between these.
-
-    // 11.1.1. Add
-    // struct {
-    //     KeyPackage key_package;
-    // } Add;
-    pub(crate) fn create_add_proposal(
-        &self,
-        framing_parameters: FramingParameters,
-        joiner_key_package: KeyPackage,
-        signer: &impl Signer,
-    ) -> Result<AuthenticatedContent, CreateAddProposalError> {
-        if let Some(required_capabilities) = self.required_capabilities() {
-            joiner_key_package
-                .leaf_node()
-                .capabilities()
-                .supports_required_capabilities(required_capabilities)?;
-        }
-        let add_proposal = AddProposal {
-            key_package: joiner_key_package,
-        };
-        let proposal = Proposal::Add(add_proposal);
-        AuthenticatedContent::member_proposal(
-            framing_parameters,
-            self.own_leaf_index(),
-            proposal,
-            self.context(),
-            signer,
-        )
-        .map_err(|e| e.into())
-    }
-
-    // 11.1.2. Update
-    // struct {
-    //     KeyPackage key_package;
-    // } Update;
-    pub(crate) fn create_update_proposal(
-        &self,
-        framing_parameters: FramingParameters,
-        // XXX: There's no need to own this. The [`UpdateProposal`] should
-        //      operate on a reference to make this more efficient.
-        leaf_node: LeafNode,
-        signer: &impl Signer,
-    ) -> Result<AuthenticatedContent, LibraryError> {
-        let update_proposal = UpdateProposal { leaf_node };
-        let proposal = Proposal::Update(update_proposal);
-        AuthenticatedContent::member_proposal(
-            framing_parameters,
-            self.own_leaf_index(),
-            proposal,
-            self.context(),
-            signer,
-        )
-    }
-
-    // 11.1.3. Remove
-    // struct {
-    //     KeyPackageRef removed;
-    // } Remove;
-    pub(crate) fn create_remove_proposal(
-        &self,
-        framing_parameters: FramingParameters,
-        removed: LeafNodeIndex,
-        signer: &impl Signer,
-    ) -> Result<AuthenticatedContent, ValidationError> {
-        if self.public_group().leaf(removed).is_none() {
-            return Err(ValidationError::UnknownMember);
-        }
-        let remove_proposal = RemoveProposal { removed };
-        let proposal = Proposal::Remove(remove_proposal);
-        AuthenticatedContent::member_proposal(
-            framing_parameters,
-            self.own_leaf_index(),
-            proposal,
-            self.context(),
-            signer,
-        )
-        .map_err(ValidationError::LibraryError)
-    }
-
-    // 11.1.4. PreSharedKey
-    // struct {
-    //     PreSharedKeyID psk;
-    // } PreSharedKey;
-    // TODO: #751
-    pub(crate) fn create_presharedkey_proposal(
-        &self,
-        framing_parameters: FramingParameters,
-        psk: PreSharedKeyId,
-        signer: &impl Signer,
-    ) -> Result<AuthenticatedContent, LibraryError> {
-        let presharedkey_proposal = PreSharedKeyProposal::new(psk);
-        let proposal = Proposal::PreSharedKey(presharedkey_proposal);
-        AuthenticatedContent::member_proposal(
-            framing_parameters,
-            self.own_leaf_index(),
-            proposal,
-            self.context(),
-            signer,
-        )
-    }
-
-    pub(crate) fn create_custom_proposal(
-        &self,
-        framing_parameters: FramingParameters,
-        custom_proposal: CustomProposal,
-        signer: &impl Signer,
-    ) -> Result<AuthenticatedContent, LibraryError> {
-        let proposal = Proposal::Custom(custom_proposal);
-        AuthenticatedContent::member_proposal(
-            framing_parameters,
-            self.own_leaf_index(),
-            proposal,
-            self.context(),
-            signer,
-        )
-    }
-}
