--- openmls/src/group/core_group/new_from_welcome.rs
@@ -1,8 +1,0 @@
-use log::debug;
-
-use crate::{
-    group::{core_group::*, errors::WelcomeError},
-    schedule::psk::store::ResumptionPskStore,
-    storage::OpenMlsProvider,
-    treesync::errors::{DerivePathError, PublicTreeError},
-};
