+++ openmls/src/group/mls_group/proposal.rs
@@ -1,26 +1,25 @@
 use openmls_traits::{signatures::Signer, storage::StorageProvider as _, types::Ciphersuite};
 
 use super::{
+    create_commit_params::CreateCommitParams,
     errors::{ProposalError, ProposeAddMemberError, ProposeRemoveMemberError, RemoveProposalError},
-    CreateGroupContextExtProposalError, CustomProposal, GroupContextExtensionProposal, MlsGroup,
-    MlsGroupState, PendingCommitState, Proposal,
+    AddProposal, CreateGroupContextExtProposalError, CustomProposal, FramingParameters,
+    GroupContextExtensionProposal, MlsGroup, MlsGroupState, PendingCommitState,
+    PreSharedKeyProposal, Proposal, QueuedProposal, RemoveProposal, UpdateProposal,
 };
 use crate::{
     binary_tree::LeafNodeIndex,
     ciphersuite::hash_ref::ProposalRef,
     credentials::Credential,
     extensions::Extensions,
-    framing::MlsMessageOut,
-    group::{
-        core_group::create_commit_params::CreateCommitParams, errors::CreateAddProposalError,
-        GroupId, QueuedProposal,
-    },
+    framing::{mls_auth_content::AuthenticatedContent, MlsMessageOut},
+    group::{errors::CreateAddProposalError, GroupId, ValidationError},
     key_packages::KeyPackage,
     messages::{group_info::GroupInfo, proposals::ProposalOrRefType},
     prelude::LibraryError,
     schedule::PreSharedKeyId,
     storage::{OpenMlsProvider, StorageProvider},
-    treesync::LeafNodeParameters,
+    treesync::{LeafNode, LeafNodeParameters},
     versions::ProtocolVersion,
 };
 
@@ -453,5 +452,124 @@ impl MlsGroup {
         self.proposal_store_mut()
             .remove(proposal_ref)
             .ok_or(RemoveProposalError::ProposalNotFound)
+    }
+
+    // === Create handshake messages ===
+    // TODO: share functionality between these.
+
+    // 11.1.1. Add
+    // struct {
+    //     KeyPackage key_package;
+    // } Add;
+    pub(crate) fn create_add_proposal(
+        &self,
+        framing_parameters: FramingParameters,
+        joiner_key_package: KeyPackage,
+        signer: &impl Signer,
+    ) -> Result<AuthenticatedContent, CreateAddProposalError> {
+        if let Some(required_capabilities) = self.required_capabilities() {
+            joiner_key_package
+                .leaf_node()
+                .capabilities()
+                .supports_required_capabilities(required_capabilities)?;
+        }
+        let add_proposal = AddProposal {
+            key_package: joiner_key_package,
+        };
+        let proposal = Proposal::Add(add_proposal);
+        AuthenticatedContent::member_proposal(
+            framing_parameters,
+            self.own_leaf_index(),
+            proposal,
+            self.context(),
+            signer,
+        )
+        .map_err(|e| e.into())
+    }
+
+    // 11.1.2. Update
+    // struct {
+    //     KeyPackage key_package;
+    // } Update;
+    pub(crate) fn create_update_proposal(
+        &self,
+        framing_parameters: FramingParameters,
+        // XXX: There's no need to own this. The [`UpdateProposal`] should
+        //      operate on a reference to make this more efficient.
+        leaf_node: LeafNode,
+        signer: &impl Signer,
+    ) -> Result<AuthenticatedContent, LibraryError> {
+        let update_proposal = UpdateProposal { leaf_node };
+        let proposal = Proposal::Update(update_proposal);
+        AuthenticatedContent::member_proposal(
+            framing_parameters,
+            self.own_leaf_index(),
+            proposal,
+            self.context(),
+            signer,
+        )
+    }
+
+    // 11.1.3. Remove
+    // struct {
+    //     KeyPackageRef removed;
+    // } Remove;
+    pub(crate) fn create_remove_proposal(
+        &self,
+        framing_parameters: FramingParameters,
+        removed: LeafNodeIndex,
+        signer: &impl Signer,
+    ) -> Result<AuthenticatedContent, ValidationError> {
+        if self.public_group().leaf(removed).is_none() {
+            return Err(ValidationError::UnknownMember);
+        }
+        let remove_proposal = RemoveProposal { removed };
+        let proposal = Proposal::Remove(remove_proposal);
+        AuthenticatedContent::member_proposal(
+            framing_parameters,
+            self.own_leaf_index(),
+            proposal,
+            self.context(),
+            signer,
+        )
+        .map_err(ValidationError::LibraryError)
+    }
+
+    // 11.1.4. PreSharedKey
+    // struct {
+    //     PreSharedKeyID psk;
+    // } PreSharedKey;
+    // TODO: #751
+    pub(crate) fn create_presharedkey_proposal(
+        &self,
+        framing_parameters: FramingParameters,
+        psk: PreSharedKeyId,
+        signer: &impl Signer,
+    ) -> Result<AuthenticatedContent, LibraryError> {
+        let presharedkey_proposal = PreSharedKeyProposal::new(psk);
+        let proposal = Proposal::PreSharedKey(presharedkey_proposal);
+        AuthenticatedContent::member_proposal(
+            framing_parameters,
+            self.own_leaf_index(),
+            proposal,
+            self.context(),
+            signer,
+        )
+    }
+
+    pub(crate) fn create_custom_proposal(
+        &self,
+        framing_parameters: FramingParameters,
+        custom_proposal: CustomProposal,
+        signer: &impl Signer,
+    ) -> Result<AuthenticatedContent, LibraryError> {
+        let proposal = Proposal::Custom(custom_proposal);
+        AuthenticatedContent::member_proposal(
+            framing_parameters,
+            self.own_leaf_index(),
+            proposal,
+            self.context(),
+            signer,
+        )
     }
 }
