+++ openmls/src/group/mls_group/staged_commit.rs
@@ -1,16 +1,17 @@
 use core::fmt::Debug;
 use std::mem;
 
+use proposal_store::{
+    QueuedAddProposal, QueuedPskProposal, QueuedRemoveProposal, QueuedUpdateProposal,
+};
+
 use super::{super::errors::*, *};
 use crate::{
     ciphersuite::{hash_ref::ProposalRef, Secret},
     framing::mls_auth_content::AuthenticatedContent,
-    group::{
-        public_group::{
-            diff::{apply_proposals::ApplyProposalsValues, StagedPublicGroupDiff},
-            staged_commit::PublicStagedCommitState,
-        },
-        QueuedAddProposal, QueuedPskProposal, QueuedRemoveProposal, QueuedUpdateProposal,
+    group::public_group::{
+        diff::{apply_proposals::ApplyProposalsValues, StagedPublicGroupDiff},
+        staged_commit::PublicStagedCommitState,
     },
     schedule::{CommitSecret, EpochAuthenticator, EpochSecrets, InitSecret, PreSharedKeyId},
     treesync::node::encryption_keys::EncryptionKeyPair,
@@ -98,7 +99,7 @@ impl MlsGroup {
     ///  - Verifies the confirmation tag
     ///
     /// Returns a [StagedCommit] that can be inspected and later merged into the
-    /// group state with [CoreGroup::merge_commit()] This function does the
+    /// group state with [MlsGroup::merge_commit()] This function does the
     /// following checks:
     ///  - ValSem101
     ///  - ValSem102
