+++ openmls/src/group/mls_group/tests_and_kats/tests/mod.rs
@@ -1,6 +1,5 @@
 //! Test and Known Answer Test (KAT) modules for the MLS group.
 
-mod core_group;
 mod create_commit_params;
 mod external_init;
 mod mls_group;
