+++ openmls/src/group/tests_and_kats/mod.rs
@@ -2,7 +2,5 @@
 mod kats;
 #[cfg(test)]
 mod tests;
-#[cfg(any(feature = "test-utils", test))]
-pub(crate) mod tree_printing;
 #[cfg(test)]
 pub(crate) mod utils;
