+++ openmls/src/prelude.rs
@@ -2,7 +2,7 @@
 //! Include this to get access to all the public functions of OpenMLS.
 
 // MlsGroup
-pub use crate::group::{core_group::Member, *};
+pub use crate::group::{Member, *};
 
 pub use crate::group::public_group::{errors::*, PublicGroup};
